[REFACTOR] CommonDto packaging 완료 #98
