+++ src/main/java/umc/tickettaka/service/TicketQueryService.java
@@ -3,7 +3,7 @@ package umc.tickettaka.service;
 import java.util.List;
 import umc.tickettaka.domain.Member;
 import umc.tickettaka.domain.ticket.Ticket;
-import umc.tickettaka.web.dto.response.CommonTicketDto;
+import umc.tickettaka.web.dto.common.CommonTicketDto;
 
 public interface TicketQueryService {
